+++ scripts/linux/install.sh
@@ -1,8 +1,4 @@
 #!/bin/bash
 
-# Dependencies
-nuget install Src/Penneo/packages.config -o Src/packages
-nuget install Src/PenneoTests/packages.config -o Src/packages
-
-# Compile
+dotnet restore Src/Penneo.sln
 dotnet msbuild Src/Penneo.sln