+++ scripts/linux/runtests.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-nunit-console Src/PenneoTests/bin/Debug/PenneoTests.dll
+dotnet vstest Src/PenneoTests/bin/Debug/netcoreapp3.1/PenneoTests.dll